Apply teacher remarks
