--- link-parser/src/main/java/ru/tinkoff/edu/java/url/LinkMatcher.java
@@ -1,5 +1,0 @@
-package ru.tinkoff.edu.java.url;
-
-public interface LinkMatcher {
-    Link matchLink(String url);
-}
