+++ link-parser/src/main/java/ru/tinkoff/edu/java/url/PatternLinkMatcher.java
@@ -1,5 +1,8 @@
 package ru.tinkoff.edu.java.url;
 
+import ru.tinkoff.edu.java.url.link.Link;
+import ru.tinkoff.edu.java.url.link.matcher.LinkMatcher;
+
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
