+++ link-parser/src/main/java/ru/tinkoff/edu/java/url/link/GitHubLink.java
@@ -1,0 +1,8 @@
+package ru.tinkoff.edu.java.url.link;
+
+public record GitHubLink(String user, String repo) implements Link {
+    @Override
+    public String getAsString() {
+        return String.format("https://github.com/%s/%s/", user, repo);
+    }
+}
