+++ link-parser/src/main/java/ru/tinkoff/edu/java/url/link/Link.java
@@ -1,4 +1,4 @@
-package ru.tinkoff.edu.java.url;
+package ru.tinkoff.edu.java.url.link;
 
 public interface Link {
     String getAsString();
