+++ link-parser/src/main/java/ru/tinkoff/edu/java/url/link/StackOverflowLink.java
@@ -1,0 +1,9 @@
+package ru.tinkoff.edu.java.url.link;
+
+public record StackOverflowLink(String id) implements Link {
+
+    @Override
+    public String getAsString() {
+        return String.format("https://stackoverflow.com/questions/%s", id);
+    }
+}
