+++ link-parser/src/main/java/ru/tinkoff/edu/java/url/link/matcher/GitHubLinkMatcher.java
@@ -1,0 +1,21 @@
+package ru.tinkoff.edu.java.url.link.matcher;
+
+import ru.tinkoff.edu.java.url.PatternLinkMatcher;
+import ru.tinkoff.edu.java.url.link.GitHubLink;
+import ru.tinkoff.edu.java.url.link.Link;
+
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+
+public class GitHubLinkMatcher extends PatternLinkMatcher {
+    private static final Pattern pattern = Pattern.compile("https://github.com/([a-zA-Z0-9_\\-]+)/([a-zA-Z0-9_\\-]+)(?:/.*)?");
+
+    public GitHubLinkMatcher() {
+        super(pattern);
+    }
+
+    @Override
+    public Link parseMatchedLink(Matcher matcher) {
+        return new GitHubLink(matcher.group(1), matcher.group(2));
+    }
+}
