+++ link-parser/src/main/java/ru/tinkoff/edu/java/url/link/matcher/LinkMatcher.java
@@ -1,0 +1,7 @@
+package ru.tinkoff.edu.java.url.link.matcher;
+
+import ru.tinkoff.edu.java.url.link.Link;
+
+public interface LinkMatcher {
+    Link matchLink(String url);
+}
